+++ CHANGELOG.md
@@ -7,12 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## Changed
+### Added
+- Filter and sort get_thread and get_thread/messages.
+
+### Changed
 - Prettier automatic tailwind classes sorting.
 - Nicer HIL validation window.
 
 ## Added
 - [frontend] - tool thunmbnail in chat.
+
+### Removed
+- `order` column in `messages` table + fixed.
+- SQLite support.
+
+## Added
+- Basic Guardrails.
+- Possibility to interrupt streaming.
 
 ## [v0.4.6]